Send mail add remove

+++ app/controllers/supervisors/trainees_controller.rb
@@ -10,27 +10,36 @@ class Supervisors::TraineesController < Supervisors::SupervisorsController
     column = [:user_id, :course_id]
     ActiveRecord::Base.transaction do
       CourseUser.import column, course_user_data
-      #AssignTraineeJob.set(wait: 10.seconds).perform_later(list_emails,
-        #params["course_user"]["course_id"], 1)
-      flash[:success] = t "messages.destroy_success"
+      AssignTraineeJob.perform_later(list_emails,
+        params["course_user"]["course_id"], 1)
+      flash[:success] = t "messages.save_success"
       redirect_to request.referrer
     end
     rescue StandardError
-      flash[:danger] = t "messages.destroy_error"
-    redirect_to request.referrer
+      flash[:danger] = t "messages.save_unsuccess"
+      redirect_to request.referrer
   end
 
   def destroy
-    if @course_user.destroy
-      msg = {status: "success", message: t("messages.destroy_success")}
-      #AssignTraineeJob.set(wait: 10.seconds).perform_later(params[:id],
-        #params[:course_id], 2)
-    else
-      msg = {status: "fail", message: t("messages.error")}
+    ActiveRecord::Base.transaction do
+      check = User.select(:email).find_by id: params[:id]
+      if check.present?
+        if @course_user.destroy
+          msg = {status: "success", message: t("messages.destroy_success")}
+          AssignTraineeJob.perform_later(check.email, params[:course_id], 2)
+        else
+          msg = {status: "fail", message: t("messages.error")}
+        end
+      else
+        msg = {status: "not_found", message: t("messages.not_found")}
+      end
+      respond_to do |format|
+        format.json{render json: msg}
+      end
     end
-    respond_to do |format|
-      format.json{render json: msg}
-    end
+    rescue StandardError
+      flash[:danger] = t "messages.destroy_error"
+      redirect_to request.referrer
   end
 
   private
@@ -44,24 +53,17 @@ class Supervisors::TraineesController < Supervisors::SupervisorsController
       params[:course_id])
     return if @course_user
     respond_to do |format|
-      msg = {status: "ok", message: t"messages.not_found"}
+      msg = {status: "ok", message: t("messages.not_found")}
       format.json{render json: msg}
     end
   end
 
   def course_user_data
-    @data = []
-    course_user_prams["user_ids"].each do |user_id|
-      @data << {user_id: user_id, course_id: course_user_prams["course_id"]}
-    end
-    @data
+    @data = course_user_prams["user_ids"].map{ |user_id|
+      {user_id: user_id, course_id: course_user_prams["course_id"]} }
   end
 
   def list_emails
-    @emails = []
-    course_user_data.each do |id|
-      @emails << User.byId(id[:user_id]).pluck(:email, :fullname)[0]
-    end
-    @emails
+    @emails = course_user_data.map { |id| User.byId(id[:user_id]).pluck(:email, :fullname)[0] }
   end
 end
